requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
asyncio-mqtt==0.16.1
python-dotenv==1.0.0